+++ src/features/auth/components/index.ts
@@ -1,7 +1,7 @@
 export * from "./Beta";
-export * from "./CreateFarm";
+export * from "./NoFarm";
 export * from "./CreatingFarm";
-export * from "./Donation";
+export * from "./CreateFarm";
 export * from "./Loading";
 export * from "./Splash";
 export * from "./StartFarm";
