+++ src/features/game/lib/gameMachine.ts
@@ -133,7 +133,7 @@ export function startGame(authContext: Options) {
               const response = await loadSession({
                 farmId: Number(authContext.farmId),
                 sessionId: context.sessionId as string,
-                token: authContext.token as string,
+                token: authContext.rawToken as string,
               });
 
               if (!response) {
@@ -219,7 +219,7 @@ export function startGame(authContext: Options) {
               farmId: Number(authContext.farmId),
               sessionId: context.sessionId as string,
               actions: context.actions,
-              token: authContext.token as string,
+              token: authContext.rawToken as string,
               offset: context.offset,
               captcha: context.captcha,
             });
@@ -297,7 +297,7 @@ export function startGame(authContext: Options) {
                 farmId: Number(authContext.farmId),
                 sessionId: context.sessionId as string,
                 actions: context.actions,
-                token: authContext.token as string,
+                token: authContext.rawToken as string,
                 offset: context.offset,
               });
             }
@@ -305,7 +305,7 @@ export function startGame(authContext: Options) {
             const session = await mint({
               farmId: Number(authContext.farmId),
               sessionId: context.sessionId as string,
-              token: authContext.token as string,
+              token: authContext.rawToken as string,
               item: (event as MintEvent).item,
             });
 
@@ -333,7 +333,7 @@ export function startGame(authContext: Options) {
                 farmId: Number(authContext.farmId),
                 sessionId: context.sessionId as string,
                 actions: context.actions,
-                token: authContext.token as string,
+                token: authContext.rawToken as string,
                 offset: context.offset,
               });
             }
@@ -341,7 +341,7 @@ export function startGame(authContext: Options) {
             const session = await sync({
               farmId: Number(authContext.farmId),
               sessionId: context.sessionId as string,
-              token: authContext.token as string,
+              token: authContext.rawToken as string,
             });
 
             return {
@@ -373,7 +373,7 @@ export function startGame(authContext: Options) {
             const session = await withdraw({
               farmId: Number(authContext.farmId),
               sessionId: context.sessionId as string,
-              token: authContext.token as string,
+              token: authContext.rawToken as string,
               amounts,
               ids,
               sfl,
