--- src/features/game/lib/whitelist.ts
@@ -1,258 +1,0 @@
-const NETWORK = import.meta.env.VITE_NETWORK as "mainnet" | "mumbai";
-
-const WITHDRAW_WHITELIST = [
-  "0xD755984F4A5D885919451eD25e1a854daa5086C9",
-  "0xc23Ea4b3fFA70DF89874ff65759031d78e40251d",
-  "0x481a58b9385868267Ff39ab285D22234B50871eD",
-  "0xA9Fe8878e901eF014a789feC3257F72A51d4103F",
-  "0x95485461fc996a39a4f60749489b93ad2d835452",
-];
-
-// Lots of duplicates but does not really matter
-const PLAY_WHITELIST = [
-  ...WITHDRAW_WHITELIST,
-  "0x1ecb0a64f8AADdDF5AC176E890A8D418694ec666",
-  "0x43B1633AE106Ab279db9CE79bc2b051F85Ab0738",
-  "0xcC9bb698585E78355211710cc1c48fa15868a5d4",
-
-  // Testing
-  "0x66686477D5Ba3e108ba57b23E52D24eAC4172572",
-  "0x74979aCF69d007cFa1e40Edc67D7f3f5bF59972D",
-
-  // Max
-  "0xd32f25Dfa932b8064A81B8254E7997CAeBc85F97",
-
-  // Artists
-  "0x002287ec2A4232467A06a92457306157a94ad9DC",
-  "0x2b3F0A1810A19250B76290Ed9b51cb9A518B3D95",
-  "0xb8Bf9f119a18513b0ABBe302986789bC385574C3",
-  "0x05863D02B228a048d467Ac7Ba8D8D42a1802DF2b",
-  "0xAA947a259E48F986F3a4d8A511BdC730410c0607",
-  "0x60959cdA4252b2E41055c81C648B5Af4eF53aD17",
-  "0x7e919704c83fFe7F19F88E72321783b06C800904",
-  "0x0E893541f9CFdf8Db2326720F0EbCAac411Ffd1E",
-  "0x8Bcc3389261679718DB663B94e255485F81d2Ab4",
-  "0x469863310Cce058fa2B2e448Ff6B6c41057e92eD",
-
-  //Developers
-  "0xeD8906723B822368DB39401e4c40a3ff828c2906",
-  "0x60826e7d0ECA970845391F1e8db8B16B46C6CfdC",
-  "0x8d680eB6b3aD31ab065B1632bB44fa1456A26548",
-  "0x05f0E2750E1Ac47934D8e668926BFE7ADAF18319",
-  "0xad1270191afc38eB44947e37107cBca6765cabFe",
-  "0xA9355c8D53D80Baa736de0F48c1D1D8b21eb0234",
-  "0xbacae1ba4b2e591e98b158aa1bef27d16099ae0f",
-  "0x068BAb172FA62635DFB3cAbCF330EdbF63522cE4",
-  "0xD637918139D44a2DeC4Bbd24F77D8Fca355Cf53E",
-  "0x358A9bb416fe97231bAD833eb18215766F5788c2",
-
-  //Beta test
-  "0x9528CE2C1F0e4E8Eeb3CBc31fb67601C3a87c2db",
-  "0x307CED27048cCe9928C69eB7296F5BeD270b1954",
-  "0x7E8931003Be6F9BE079070A2e8738D678710F238",
-  "0xfB33c8f8426Db534C0BA6494D4B9e824B390d658",
-  "0x6b01e5C6d32d4Aa79E1B0388155EcfE7697337D7",
-  "0x5e4cb298C7ED72796D4e5e76aE7B33f3C158E809",
-  "0xD26Cd4285c5e2A61f62E91E61C4309E737955ea7",
-  "0x7666f37D4B861eccc7794eC4d86f5a86686135E9",
-  "0x7Ea3BA63370C5b90c753AEFaeE612A06547688c2",
-  "0x9701FD645b3738dd0C0Fa458bC05Bc333E60426a",
-  "0x98F6723C38eD4d49a4A374a65B98Cd9D2d47F9c2",
-  "0x90CEB8A369B7aEA5C536290fc0745081D23791f2",
-  "0xd15EFe11A5C005ccFD5b377067B54c71D421cc2B",
-  "0xb680D032EbAEBa99261b52DC26b53794578E3201",
-  "0xCee8447dd036cdF5b5aeDa4FC8F880fdE479B13D",
-  "0x4381fa0c0E162Ce7c4330b47585d921b2200a949",
-  "0x89fF8e27b92c3Cf36bEf6580d11e1921Ea6Fa93E",
-  "0x87C58AEd46c67e787761b96D55E03dCEa7e4325A",
-  "0x72d5161f427D3d08f33a65A682CdFFC2E7B59Aaa",
-  "0xfaE50Dd6eE3CA1c7491786Ad5aB30D3Ee3982Cb9",
-  "0x911E6364D7e6061fC3de135F75CAb9d61A6e5FA9",
-  "0x0E893541f9CFdf8Db2326720F0EbCAac411Ffd1E",
-  "0xe7aF00Befc5A93735710E21E8A5a6E3A3888bc7c",
-  "0xF3Ef97062FB59A8e6c460D9F267a162df4C20843",
-  "0x60959cdA4252b2E41055c81C648B5Af4eF53aD17",
-  "0x05863D02B228a048d467Ac7Ba8D8D42a1802DF2b",
-  "0xC297434911D5A22eeF2925a63E78D211A77C31C4",
-  "0x7666f37D4B861eccc7794eC4d86f5a86686135E9",
-  "0xad1270191afc38eB44947e37107cBca6765cabFe",
-  "0x5e878c5D016c123856ecFAeD7FA9625B39fF7c3c",
-  "0xCee8447dd036cdF5b5aeDa4FC8F880fdE479B13D",
-  "0x068BAb172FA62635DFB3cAbCF330EdbF63522cE4",
-  "0xBe469243177E4040F5C659B682733ABd2226d156",
-  "0xBe469243177E4040F5C659B682733ABd2226d156",
-  "0x832877b07a67EbE54f7c83E6cD561aB1e25bDa01",
-  "0xEE66195f7438012F8F3Dae456c090a794Af55CBF",
-  "0x654dd1ec75ce660899383f627882c7f3dfa8ffd0",
-  "0xfB33c8f8426Db534C0BA6494D4B9e824B390d658",
-  "0xf89941FEF5E65CA7337D9165F03304044C0612a8",
-  "0x98F6723C38eD4d49a4A374a65B98Cd9D2d47F9c2",
-  "0x2FC519190428c982D43Bbe1e7df9F639a91bC1d9",
-  "0x654dd1ec75ce660899383f627882c7f3dfa8ffd0",
-  "0xd15EFe11A5C005ccFD5b377067B54c71D421cc2B",
-  "0x040325b8bbAF8e894D4dC34B28415cD48C34Cb83",
-  "0x14e7502250De7297aB29aB5142C5AB0a460b03CA",
-  "0x2b3F0A1810A19250B76290Ed9b51cb9A518B3D95",
-  "0x0eC3D880fD1E0107bdaE95430B199add7E2F27fD",
-  "0xC895a83f22CFff38065b998968c2e2Ef4EB18413",
-  "0xa5967f7e0a732263b1561eC888f0387d5F7b7789",
-  "0x424BEb94dF6bd826a2e92a3C2C44C35257792d24",
-  "0xfaE50Dd6eE3CA1c7491786Ad5aB30D3Ee3982Cb9",
-  "0xda816828d3606F98f2187011ca7ceFD7fE4d2868",
-  "0x170631B2bC1C893cDEb274683f2999eF98437221",
-  "0x5D713Cf527e65c7aA6EB83AaA0f399e09Ffcd99a",
-  "0xd9b75d557821C5233fa6F4b78c936680263e8DA5",
-  "0xec2e7EE9437e3Bc0C0E94aaF573C3910Fa66160a",
-  "0x3048e95798Fac87D8eF240Da91Ce5EA7F7eF9e80",
-  "0x9f920a8c3307625AAE5FD7E6647Ae8a15D959092",
-  "0xC5E3868876F43a72Ee7492C6ed3538b47E97f6dd",
-  "0x4ca587D06FfA3D3FE2D83E6690cE228374906879",
-  "0xf7f71De0e7139De66fD0603b402EeCd6B3DaeAE3",
-  "0x4D24f3Fce02c19A8A776a404246ea5E1537EA553",
-  "0x53cF3645c491241B35c0a90ebfdE5b91A6bFc763",
-  "0xCEEEF01ef73FfA53D28a1A0fD85Fc602B5E60bd0",
-  "0x124f18f7001780dd4DCcc73BA649e926468de065",
-  "0xAeb17aE1a481be04380b321Fb464e9EeD8615Aaa",
-  "0x7898A866B011A1761639E14d93cFd72D04d1Ddef",
-  "0x026B2C1D3056ab0660e9416537feD07636A9E9Bc",
-  "0x47311fcD99cfb2F0764054b2546cBE0467d98fD2",
-  "0xd19677a52e2517751F1db3F1341D31FFa0066144",
-  "0xeaB466d01f673c200408c32291c15DDCE61aE928",
-  "0x071dfb36347B3Def1069bFC36cEFF0557e385e1a",
-  "0xE13A7e397697F9C2A717558b7F0d74Db54A88E14",
-  "0x62003019B2c71BA5C1B0Bcf86246B5ec1E8B8952",
-  "0x1CBb844ff59E66C23BAf7b2B86eb49105B054979",
-  "0x264EBdd59F1a46C9eDEdFeD174A1Eb140fb5e49a",
-  "0xCaDcc0f0f7C1FcC121bf2769da2d5EF0FCC0ECAA",
-  "0x00D27Df1E552C47E5F7ab62DC4138f977565Ec84",
-  "0x297Dfb66753dd2be08f977DeAb31f675418364b9",
-  "0x6435071C47A5F1dd9Ba225A48fa4381BE928Df17",
-  "0x5f5801a905503B1AEDE5E93924eB83746bFB2F1b",
-  "0xc84F151276b93a9c94B7642c738C9e5c4Cf6a316",
-  "0x48b9f15128cca0d092F254968FAbA2265dE6f0AF",
-  "0xdBF605dF6999Ed2d65985913f11B3075b30ccD8c",
-  "0x2aC7DaA7D494E55B5717684BE8da5C957b1692Bd",
-  "0x93c1ff68200506cE5e7E1707968F0ba9Cf76c433",
-  "0xDb457e5889f4d4c7ed2F443fF4003754582cf7FC",
-  "0x14656c7B1DB495f9efca0B6770d8306297FEe34A",
-
-  "0x25F733235028c96d758E2D0332B340C3B920a7B8",
-  "0x3B5A583C1B0d0A85a592d2060Ff57da2cB754eB3",
-  "0x654DD1Ec75ce660899383f627882C7f3dfa8Ffd0",
-  "0xB2F5c580A0Ba371a31Ab2fb0DF69578F8c6f4DAd",
-  "0x91b4683F7d937C56b723E0a9Da8bAfBEa6d82371",
-  "0xb8Bf9f119a18513b0ABBe302986789bC385574C3",
-  "0x654DD1Ec75ce660899383f627882C7f3dfa8Ffd0",
-  "0xB2F5c580A0Ba371a31Ab2fb0DF69578F8c6f4DAd",
-  "0xc147d7E53c535576712a6d488578C8C6E1d72954",
-  "0x8fc5fb7E5D62D8CAf0a298dC5b89df6061d8F183",
-  "0x0c274631545911CaF8E709917c571E69eDa2E6eE",
-  "0x5e678BF568e031ae3Cde87D6BFdd9eC3e9aDaBdC",
-  "0x29685Af51FE8D0782BEB44F20151C9afF52Aaf5D",
-  "0x85eFBd881B4dd0bF538b2E4352cdAEe29E082f01",
-  "0x68Fbf5C5D0dA6f65b719188E463e33261eabd62F",
-  "0xC488D0858d5d09250544F9141B0c507b4a94c37c",
-  "0xFFC2aa713e364baBAAfAe397697693570ff22762",
-  "0x92Db204F468B8a36428cD086C3c7B1512d70dF51",
-  "0x968A33Fb1580Fd250C797997D281b7dE9DbFeAE0",
-  "0x3d64F3b693BcBCAeeeF5DECF0E6402F25Fe37790",
-  "0xb9a90580c5ae5511fDb6cbA0785eFDFcC6d0e5E2",
-  "0x8b9006503C7fc05A677d2612d7931DF8f898f6c3",
-  "0x08317190d8D942C9E9CC55DAA49E46E3c11f4650",
-  "0x99827F0D6A099237Dc654d05A6B919920Adc3A46",
-  "0x04d8B1aD508661FeB1C4fa4EFD78621458c7B2EE",
-  "0x7e07dAD511db83Fce05f99dB2dD0e9a91c540CA1",
-  "0x766022f44BED48EF670160f61b051B89A1182dCE",
-  "0x0e85B66196A1F639BE68B012cd1E4ca516BfaD24",
-  "0xf9bBbc96FF924aD33b3699624387Ec6E488Ff92C",
-  "0xf5d5A404827f272EAb6C7BEf3eBB633A5796EB64",
-  "0x469863310Cce058fa2B2e448Ff6B6c41057e92eD",
-  "0x20cf434aE3Ff233080696f626e481027C1D9f839",
-  "0x5a57107A58A0447066C376b211059352B617c3BA",
-  "0xA9cB1c0ADfc9655641A0d022be3A71a846Bb24C7",
-  "0xac4f94Ff8c14dBd35abe97416EeA8De08dc119a1",
-  "0x002287ec2A4232467A06a92457306157a94ad9DC",
-  "0x5b4Ea42B3Bad0F7eCc4c60121905D8180233a08E",
-  "0x90AeB4691f438f42675F1feB611aFe8B4c40a2db",
-  "0x8A6F4372cF69c69a16e6fd2f9bd379012dE79242",
-  "0x7176e3e880dfD4596a0A0A35AE4fE25f458C9D4e",
-  "0xbe5476F7a2E56645530a147fE50652e377BefC74",
-  "0xaa53FF3118c85A6df5965a89b41F060ef593D8fb",
-  "0x69061ABF1bb7fd7540AC8AE4BD624589BFd0c814",
-  "0xfB4B3Bb612dCFC53CAb538e29Ff5A64d3868509a",
-  "0x96f9F7Ca06E3Eb2825E55474eF92e6549AA3a5F6",
-  "0xbc0054143C9c3D3f3a4ADcaa45283E831105a4cc",
-  "0x1A28330434582a460240C2748092BDaa395Dd765",
-  "0xe1e91A1de232367855397DB4B63447029331c5a8",
-  "0xC1da309F85eD397A942f9bF212cdd63110E61515",
-  "0x6934E19eD160984537D0d8dc0Bf4efFF81297284",
-  "0x131dC6E3Ea0ECd24b83100C96fb06885aBFf8d31",
-  "0xb427FF96D9B59dd90BC42c2DDeA08830D67dEE89",
-  "0x52aa15875960Bf99A0338882FA7e098ccA5F7d0b",
-  "0xb0c36Cdc1e9BFF41049ed0f60270a19850BEF1b3",
-  "0xD28a88942d0d289e834ada6E726f28Fd245e026b",
-  "0x720aFB65fCBE593E5e85CbA8d45aC8f94d1CD7D7",
-  "0x23d2B5a96528B63dbcFe49F8CEB11Cbe27F6D527",
-  "0x3e93111efCf58d50743F98785e1761C9639B8177",
-  "0xD57c3c4c3CF9915EB9680cA147203CD3df6BF0bf",
-  "0x9db76D0020E9446E87947FA38DE3e3276c83867d",
-  "0x0289b123794799C9fDD42f49D77AfC2cbFF0593f",
-  "0xD3A68B6B2e184Ce1b325C9f811DbDCcF975eCab4",
-  "0x49666CF3B2A8293da1CA2c5897e99fd7CCD3283C",
-  "0x5AF9169deaD147d3D963057d99a910a0B16DC025",
-  "0x4DC84Bfd85B4c6645f7E4Eb86c615Ce8a7290Fa7",
-  "0x2b1eb5D1FB443d872e7ca3A82E295BC8080cD403",
-  "0x03F4B156e8C077219A69A4EC63fAcA4B449b7723",
-  "0xAe0B1282F57962929C498f5383EAeeE7a93dD147",
-  "0xB2285bC9703b4B20A13D9960ad93E10cfF9C3e4c",
-  "0xe423661b92866c712f6c7516bd9fbc52c202eacb",
-  "0xB58b7E5Deb295873E088889Df6d92f2B55a0ab94",
-  "0x9d6fC525bd7f8F8f2759E747F29BfCA9d5bf5ff1",
-  "0x45fCbE33B00BA4265C6E4B814fB5f060BF2cD97b",
-  "0x97E18713C64F75b5c4c6B51b7AB4495c357Db543",
-  "0x60826e7d0ECA970845391F1e8db8B16B46C6CfdC",
-  "0xc6Bd19086b02522a8AE2606194052Af46770717e",
-  "0x97E18713C64F75b5c4c6B51b7AB4495c357Db543",
-  "0x13402fDD4B6554d3fCC1aDdfe28CE68D98Dac145",
-  "0x89c18aBfF65d79c9D8711c47927Ac1Dc64610099",
-  "0x05980ad62b176dEC24AfEB44FC0a2543942C7559",
-  "0x23d2B5a96528B63dbcFe49F8CEB11Cbe27F6D527 ",
-  "0x6d9a4004C80c575a1d1BD8DaEfdf609CE436cA6D",
-  "0x41ca85eFcd4Df1599C97F40cbE5A64A412baEE5D",
-  "0x449B3dA0b9799867bd76AF28A77944EC275A6069",
-  "0x7197c5F15a61bA5e6101e7C02491B33a47D36923",
-  "0x4ca790b2C94Af33009F5E7E2A419d766d77AA88A",
-  "0x4ca790b2C94Af33009F5E7E2A419d766d77AA88A",
-  "0x7D3b9c8D0f2FF463959469001142D185a992d387",
-  "0xDfC70F5205a2bC1bD44219fd852D720d7a36c5b8",
-  "0xCCa91FDAD205214aE5005082De56330538C99e09",
-  "0x091E577FC0B068a61D39604Db86faE809D632A90",
-  "0x3E415cBd89D9C5F0e7476E0F3E7dfe984d0f9Fef",
-];
-export function canCreateFarm(address: string) {
-  if (NETWORK !== "mainnet") {
-    return true;
-  }
-  return PLAY_WHITELIST.includes(address);
-}
-
-export function canSync(address: string) {
-  if (NETWORK !== "mainnet") {
-    return true;
-  }
-  return PLAY_WHITELIST.includes(address);
-}
-
-export function canMint(address: string) {
-  if (NETWORK !== "mainnet") {
-    return true;
-  }
-  return WITHDRAW_WHITELIST.includes(address);
-}
-
-export function canWithdraw(address: string) {
-  if (NETWORK !== "mainnet") {
-    return true;
-  }
-  return WITHDRAW_WHITELIST.includes(address);
-}
