+++ src/features/hud/components/Menu.tsx
@@ -16,11 +16,9 @@ import mobileMenu from "assets/icons/hamburger_menu.png";
 import questionMark from "assets/icons/expression_confused.png";
 import radish from "assets/icons/radish.png";
 import water from "assets/icons/expression_working.png";
-import token from "assets/icons/token.gif";
 import timer from "assets/icons/timer.png";
 import { useTour } from "@reactour/tour";
 import { TourStep } from "features/game/lib/Tour";
-import { canSync } from "features/game/lib/whitelist";
 import { metamask } from "lib/blockchain/metamask";
 
 export const Menu = () => {
@@ -94,7 +92,7 @@ export const Menu = () => {
   }, []);
 
   const syncOnChain = async () => {
-    if (!canSync(metamask.myAccount as string)) {
+    if (!authState.context.token?.userAccess.sync) {
       setShowComingSoon(true);
       return;
     }
