+++ src/lib/config.ts
@@ -11,12 +11,14 @@ const INVENTORY_CONTRACT = import.meta.env.VITE_INVENTORY_CONTRACT;
 const PAIR_CONTRACT = import.meta.env.VITE_PAIR_CONTRACT;
 const SESSION_CONTRACT = import.meta.env.VITE_SESSION_CONTRACT;
 const TOKEN_CONTRACT = import.meta.env.VITE_TOKEN_CONTRACT;
+const DISCORD_REDIRECT = import.meta.env.VITE_DISCORD_REDIRECT;
 
 export const CONFIG = {
   NETWORK,
   POLYGON_CHAIN_ID,
   DONATION_ADDRESS,
   API_URL,
+  DISCORD_REDIRECT,
 
   WISHING_WELL_CONTRACT,
   BETA_CONTRACT,
